move customMerge to utils
